install: Move Aleph bits to a separate module
install.rs is getting way too large.
